+++ CHANGELOG.md
@@ -23,6 +23,7 @@ server
 * fixed image label name for url context from `url.server` to `url.context`; this fixes the url context not being shown in urls 
   (thanks to Andreas K)
 * added `ENGINE_LOG` server configuration
+* fixed empty stage validation emails (thanks to Sebastian D)
 * fixed container prune before startup: the previous version only purged containers if the images didn't change
   (also fixes images not being removed by new builds)
 * start api command: fixed rare https argument problem